ckd/requirements.txt
absl-py==1.0.0
astor==0.8.1
cached-property==1.5.2
cachetools==5.0.0
certifi==2021.10.8
chardet==4.0.0
click==8.1.2
colorama==0.4.4
Flask==2.1.1
gast==0.5.3
google-auth==2.6.6
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.1
Jinja2==3.1.1
joblib==1.1.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.6
MarkupSafe==2.1.1
numpy==1.22.3
oauthlib==3.2.0
opt-einsum==3.3.0
pickle-mixin==1.0.2
Pillow==9.1.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
tensorboard==2.8.0
tensorflow==2.8.0
tensorflow-estimator==2.8.0
termcolor==1.1.0
typing-extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.1
wrapt==1.14.0
zipp==3.8.0
